# contracts/openapi/topsdojob-v3-local.yaml
openapi: 3.1.0
info:
  title: Tops do Job V3 Local API
  version: 0.0.1
  description: "Contrato minimo local do skeleton, camada transversal, API publica de leitura e funil local Anuncie gratis. A API segue fail-closed: endpoints nao descritos em /api/** nao devem ficar publicos por acidente."
servers:
  - url: http://localhost:8080
    description: backend local
paths:
...
  /api/admin/anuncios/resumo:
    get:
      summary: Resumo administrativo local de anuncios
      description: Retorna apenas contagens agregadas de anuncios para ADMIN ou MODERADOR com ANUNCIO_LER. Nao retorna WhatsApp real nem dados privados.
      operationId: getAdminAnunciosResumo
      security:
        - AdminSessionCookie: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
      responses:
        "200":
          description: resumo read-only de anuncios
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminResumoAnuncios"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/admin/moderacao/resumo:
    get:
      summary: Resumo administrativo local de moderacao
      description: Retorna apenas contagens agregadas de filas de moderacao para ADMIN ou MODERADOR. Nao executa aprovacao, reprovacao ou escrita.
      operationId: getAdminModeracaoResumo
      security:
        - AdminSessionCookie: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
      responses:
        "200":
          description: resumo read-only de moderacao
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminResumoModeracao"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/admin/midias/resumo:
    get:
      summary: Resumo administrativo local de midias
      description: Retorna apenas contagens agregadas de midias e arquivos para ADMIN ou MODERADOR. Nao retorna bucket, chave de storage, hash, etag ou URL privada.
      operationId: getAdminMidiasResumo
      security:
        - AdminSessionCookie: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
      responses:
        "200":
          description: resumo read-only de midias
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminResumoMidias"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/admin/metricas/resumo:
    get:
      summary: Resumo administrativo local de metricas
      description: Retorna contagens agregadas de visualizacoes e cliques WhatsApp somente para ADMIN. Nao retorna IP, User-Agent, referer, WhatsApp real ou identificador bruto.
      operationId: getAdminMetricasResumo
      security:
        - AdminSessionCookie: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
      responses:
        "200":
          description: resumo read-only de metricas
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminResumoMetricas"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    PublicSessionCookie:
      type: apiKey
      in: cookie
      name: JSESSIONID
      description: Sessao publica emitida pelo login; cookie HttpOnly, Secure em HML e SameSite Lax.
    AdminSessionCookie:
      type: apiKey
      in: cookie
      name: JSESSIONID
      description: Sessao administrativa local emitida por `POST /api/admin/auth/login`; cliente deve usar credentials include e o corpo da resposta nunca retorna token ou cookie.
  headers:
    RequestId:
      description: Identificador opaco de correlação da requisição.
      schema:
        type: string
        example: 11111111-1111-4111-8111-111111111111
  parameters:
...
    RequestId:
      name: X-Request-Id
      in: header
      required: false
      description: Identificador opaco enviado pelo cliente para correlação.
      schema:
        type: string
        minLength: 8
        maxLength: 128
        pattern: "^[A-Za-z0-9._:-]{8,128}$"
...
  responses:
...
    Unauthorized:
      description: Autenticacao necessaria ou credenciais invalidas.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
    Forbidden:
      description: Usuario autenticado sem permissao suficiente.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
...
    InternalError:
      description: Falha inesperada sem detalhes internos.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
  schemas:
...
    AdminContador:
      type: object
      additionalProperties: false
      required:
        - codigo
        - rotulo
        - total
      properties:
        codigo:
          type: string
        rotulo:
          type: string
        total:
          type: integer
          format: int64
          minimum: 0
    AdminResumoAnuncios:
      type: object
      additionalProperties: false
      required:
        - totalAtivos
        - publicados
        - pendentesRevisao
        - pausados
        - bloqueados
        - comContatoConfigurado
        - porStatus
      properties:
        totalAtivos:
          type: integer
          format: int64
          minimum: 0
        publicados:
          type: integer
          format: int64
          minimum: 0
        pendentesRevisao:
          type: integer
          format: int64
          minimum: 0
        pausados:
          type: integer
          format: int64
          minimum: 0
        bloqueados:
          type: integer
          format: int64
          minimum: 0
        comContatoConfigurado:
          type: integer
          format: int64
          minimum: 0
          description: Contagem agregada; nao retorna telefone ou WhatsApp real.
        porStatus:
          type: array
          items:
            $ref: "#/components/schemas/AdminContador"
    AdminResumoModeracao:
      type: object
      additionalProperties: false
      required:
        - revisoesAbertas
        - revisoesEmAnalise
        - anunciosPendentesModeracao
        - anunciosBloqueados
        - documentosPendentes
      properties:
        revisoesAbertas:
          type: integer
          format: int64
          minimum: 0
        revisoesEmAnalise:
          type: integer
          format: int64
          minimum: 0
        anunciosPendentesModeracao:
          type: integer
          format: int64
          minimum: 0
        anunciosBloqueados:
          type: integer
          format: int64
          minimum: 0
        documentosPendentes:
          type: integer
          format: int64
          minimum: 0
          description: Contagem agregada; nao retorna documento privado.
    AdminResumoMidias:
      type: object
      additionalProperties: false
      required:
        - arquivosTotal
        - arquivosPendentes
        - arquivosValidados
        - midiasPublicaveis
        - midiasPendentes
        - midiasBloqueadas
        - storiesPublicados
        - storiesPendentes
      properties:
        arquivosTotal:
          type: integer
          format: int64
          minimum: 0
        arquivosPendentes:
          type: integer
          format: int64
          minimum: 0
        arquivosValidados:
          type: integer
          format: int64
          minimum: 0
        midiasPublicaveis:
          type: integer
          format: int64
          minimum: 0
        midiasPendentes:
          type: integer
          format: int64
          minimum: 0
        midiasBloqueadas:
          type: integer
          format: int64
          minimum: 0
        storiesPublicados:
          type: integer
          format: int64
          minimum: 0
        storiesPendentes:
          type: integer
          format: int64
          minimum: 0
    AdminResumoMetricas:
      type: object
      additionalProperties: false
      required:
        - visualizacoesTotal
        - cliquesWhatsappTotal
        - cliquesWhatsappPermitidos
        - cliquesWhatsappBloqueados
      properties:
        visualizacoesTotal:
          type: integer
          format: int64
          minimum: 0
        cliquesWhatsappTotal:
          type: integer
          format: int64
          minimum: 0
        cliquesWhatsappPermitidos:
          type: integer
          format: int64
          minimum: 0
        cliquesWhatsappBloqueados:
          type: integer
          format: int64
          minimum: 0
...
    ApiErrorResponse:
      type: object
      additionalProperties: false
      required:
        - timestamp
        - status
        - error
        - code
        - message
        - path
        - requestId
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          example: 500
        error:
          type: string
          example: Internal Server Error
        code:
          $ref: "#/components/schemas/ApiErrorCode"
        message:
          type: string
          example: Erro interno inesperado.
        path:
          type: string
          example: /api/health
        requestId:
          type: string
          example: 11111111-1111-4111-8111-111111111111
    ApiErrorCode:
      type: string
      enum:
        - BAD_REQUEST
        - UNAUTHORIZED
        - FORBIDDEN
        - NOT_FOUND
        - CONFLICT
        - UNPROCESSABLE_ENTITY
        - BAD_GATEWAY
        - SERVICE_UNAVAILABLE
        - INTERNAL_ERROR
...